Added note about not hammering ex from other sources.
I specifically made a callout for ex because that can also be compiled
from sources that aren't vim. In this case, I decided to call it viex,
as I don't know of another naming conflict for that. As for ed, vim
doesn't install one of those, as it's a separate package.