Removed the call to autogen.sh, as it's not in the vim tree.
This bug came about because in the emacs source tree, there is an
autogen.sh for making configure. However, vim already has a configure
script as part of the files delivered either from git or as part of
the release tarball of source code, so it doesn't need autogen.sh.