Changed (finally) the default location for vim source code.
It used to be at a location specific to my machine, and so was not
appropriate to pass that on to all the users - it's better to use
something more generic that is also under a user's direct control.
You can still change it by feeding VIMCOMPILEHOME in from the env.