Added long switches and bash-complete script.
At the moment, tangling to user's local directory seems like a bad
idea, the alternative would be to try to add it to the list of files
to install with the Makefile. I can't initially use the same filename,
or I'll clobber the original vimmaker script, so I'll have to cp it to
the destination as vimmaker, instead of bash-complete-vimmaker.