Added direct support for updating the mercurial mirror.
This presumes that the mercurial tree hasn't been extracted to the
same location as the official github clone was extracted to. If it
were, then you'll need to check for a .hg directory as well as a .git
directory, and be aware that compiles will use commits from the
mercurial repository, not the github repository.

I also haven't put _any_ support in here for compiling the mercurial
tree, so you're on your own for that for the moment until I implement
some better switch handling. At the moment, I limit it to one switch,
and any more is treated like an error, hence I can't yet change to the
mercurial repo, THEN configure/compile. That may come later.