Added note about mercurial tree.
Currently, while we fetch the mercurial tree, we don't do anything
else with the tree from this script. It would be nice to set that up
so this is a one-stop-shop.

Note for later: We should advise user that the hg tree is in a
different place than the canonical git source code for vim.